simulator/requirements.txt
pygame
pyopengl
pillow
